# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.yourdomain.com/api/v1",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.yourdomain.com/api/v1",
      "description": "Production Environment"
    }
  ],
  "info": {
    "title": "Omni-Backend API",
    "version": "1.0.0",
    "description": "Generated by nestia - https://github.com/samchon/nestia",
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "schemas": {
...
      "CreateMemoDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      },
      "UpdateMemoDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      },
      "InvitePADto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "platform": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformAccessDto"
            }
          }
        },
        "required": [
          "email",
          "firstname",
          "lastname",
          "platform"
        ]
      },
      "PlatformAccessDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "X",
              "Instagram",
              "LinkedIn",
              "Telegram",
              "Email"
            ]
          },
          "canSend": {
            "type": "boolean"
          },
          "viewMessages": {
            "type": "boolean"
          }
        },
        "required": [
          "platform",
          "canSend",
          "viewMessages"
        ]
      },
      "CollectOnboardingEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "AuthTokens": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "VerifyMfaDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "token"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TelegramConnectStep2": {
        "type": "object",
        "properties": {
          "loginId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "twoFAcode": {
            "type": "string"
          }
        },
        "required": [
          "loginId",
          "code"
        ]
      },
      "ConnectEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "imapPassword": {
            "type": "string"
          },
          "smtpPassword": {
            "type": "string"
          },
          "imapHost": {
            "type": "string"
          },
          "imapPort": {
            "type": "number"
          },
          "imapSecure": {
            "type": "boolean"
          },
          "smtpHost": {
            "type": "string"
          },
          "smtpPort": {
            "type": "number"
          },
          "smtpSecure": {
            "type": "boolean"
          }
        },
        "required": [
          "email",
          "imapPassword",
          "smtpPassword",
          "imapHost",
          "imapPort",
          "imapSecure"
        ]
      },
...
    }
  },
  "paths": {
...
    "/memo/{conversationId}": {
      "post": {
        "tags": [
          "Memo"
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "conversationId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "content",
                    "conversationId",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "memo",
          "createMemo"
        ],
        "x-samchon-controller": "MemoController"
      },
      "get": {
        "tags": [
          "Memo"
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string"
                      },
                      "conversationId": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "id",
                      "content",
                      "conversationId",
                      "createdAt",
                      "updatedAt"
                    ]
                  }
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "memo",
          "getMemos"
        ],
        "x-samchon-controller": "MemoController"
      }
    },
    "/memo/{memoId}": {
      "patch": {
        "tags": [
          "Memo"
        ],
        "parameters": [
          {
            "name": "memoId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "conversationId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "content",
                    "conversationId",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "memo",
          "updateMemo"
        ],
        "x-samchon-controller": "MemoController"
      },
      "delete": {
        "tags": [
          "Memo"
        ],
        "parameters": [
          {
            "name": "memoId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        },
        "x-samchon-accessor": [
          "memo",
          "deleteMemo"
        ],
        "x-samchon-controller": "MemoController"
      }
    },
...
    "/auth/invite-pa": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitePADto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "userId"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "invite_pa",
          "invitePa"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/request-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectOnboardingEmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "request_otp",
          "requestOtp"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "mfaEnabled": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "status",
                    "mfaEnabled"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "verify_otp",
          "verifyOtp"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/generate-2fa": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectOnboardingEmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "base64Url": {},
                    "secret": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "base64Url",
                    "secret"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "generate_2fa",
          "generate2FA"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/verify-2fa": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "verify_2fa",
          "verify2FA"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "accessToken"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "refresh"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
...
    "/auth/verify-telegram": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramConnectStep2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "session"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "verify_telegram",
          "verifyTelegram"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
...
    "/auth/x/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "username": {
                      "type": "string"
                    },
                    "externalAccountId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "username",
                    "externalAccountId"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "x",
          "callback",
          "xCallback"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/connect-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectEmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "connect_email",
          "connectEmail"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
...
  },
...
}